rename : repository 패키지명 수정
패키지명이 repository2으로 되어있어 이를 repository로 수정
