+++ src/main/java/com/woory/backend/repository/TopicRepository.java
@@ -1,6 +1,5 @@
-package com.woory.backend.repository2;
+package com.woory.backend.repository;
 
-import com.woory.backend.entity.Group;
 import com.woory.backend.entity.Topic;
 
 import org.springframework.data.jpa.repository.JpaRepository;
