+++ src/main/java/com/woory/backend/repository/TopicSetRepository.java
@@ -1,4 +1,4 @@
-package com.woory.backend.repository2;
+package com.woory.backend.repository;
 
 import java.util.Optional;
